pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "bmt-scripts"
version = "0.0.1"
description = "Python script collection for various tasks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    {name = "BMT Open Source", email = "inf@bemind.tech"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
dependencies = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
agents = [
    "pyautogen>=0.2.0",
    "python-dotenv>=1.0.0",
]
hardware = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pyautogui>=0.9.54",
]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "bandit>=1.7.0",
]
all = [
    "pyautogen>=0.2.0",
    "python-dotenv>=1.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pyautogui>=0.9.54",
]

[project.scripts]
bmtlab = "bmt_scripts.cli:main"

[project.urls]
"Bug Reports" = "https://github.com/bemindlab/bmt-scripts/issues"
"Source" = "https://github.com/bemindlab/bmt-scripts"
"Documentation" = "https://github.com/bemindlab/bmt-scripts/wiki"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["bmt_libs", "bmt_scripts"]

[tool.setuptools_scm]
write_to = "src/bmt_scripts/_version.py"
version_scheme = "post-release"
local_scheme = "node-and-date"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.bandit]
exclude_dirs = ["tests", "docs"]
targets = ["src", "scripts"]
skips = [
    "B101",  # Skip assert statements
    "B404",  # Skip subprocess import warning
    "B603",  # Skip subprocess call warning
    "B607",  # Skip subprocess partial path warning
    "B110",  # Skip try-except-pass warning
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
